requirements.txt
djangorestframework_jwt==1.11.0
Django==2.1.2
djoser==1.3.1
django_sendsms==0.3.1
twilio==6.19.2
djangorestframework==3.9.0
django_reversion==3.0.1
sendsms==0.2.0
